fetch all data from the api at once
